move SQL Server implementation examples to their own package
