+++ src/amazon_product_metadata_parser/output/SQLServer2012/SQLServer2012Output.java
@@ -1,13 +1,12 @@
-package amazon_product_metadata_parser.output;
+package amazon_product_metadata_parser.output.SQLServer2012;
 
-import java.io.Console;
-import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
-import amazon_product_metadata_parser.Parser;
 import amazon_product_metadata_parser.dto.ProductDTO;
+import amazon_product_metadata_parser.output.Output;
+import amazon_product_metadata_parser.output.ProductOutputException;
 
 /**
  * Outputs product data to the SQL Server 2012 database "AmazonData" using Windows authentication.
@@ -21,12 +20,12 @@ public class SQLServer2012Output implements Output {
   private final String serverUrl;
   private final String pcName = System.getenv(PC_NAME_WINDOWS_ENVIRONMENT_VARIABLE);
   private final String     databaseName;
-  private final JDBCStatementsToExecute statementsToExecute;
+  private final SQLServerOperations statementsToExecute;
   private       Connection conn;
   private int insertStatementsExecuted = 0;
   private int rowsAffected = 0;
 
-  public SQLServer2012Output(String databaseName, JDBCStatementsToExecute statementsToExecute) {
+  public SQLServer2012Output(String databaseName, SQLServerOperations statementsToExecute) {
     this.databaseName = databaseName;
     this.statementsToExecute = statementsToExecute;
     serverUrl = buildServerURL();
@@ -39,30 +38,6 @@ public class SQLServer2012Output implements Output {
            + ";databaseName="
            + databaseName
            + ";integratedSecurity=true";
-  }
-
-  public static void main(String... args) throws IOException {
-    System.out.printf("This program parses the 'amazon-meta.txt' file and inserts the data into "
-                      + "a SQL Server 2012 database.%nIt assumes you have set up the appropriate"
-                      + " stored procedures and schema in a database named 'AmazonProductMetadata'"
-                      + ".%n");
-    System.out.println("Press enter the path to the 'amazon-meta.txt' file to proceed or "
-                       + "press Ctrl+C to exit");
-    Console console = System.console();
-    String response = console.readLine();
-    switch (response) {
-      case "quit":
-        // Fall through
-      case "cancel":
-        // Fall through
-      case "exit":
-        break;
-      default:
-        SQLServer2012Output output = new SQLServer2012Output("AmazonProductMetadata",
-                                                             new JDBCStatementsToExecuteImplExample());
-        Parser parser = new Parser(response, output);
-        parser.parse();
-    }
   }
 
   @Override
