+++ test/amazon_product_metadata_parser/output/SQLServer2012OutputTest.java
@@ -2,6 +2,8 @@ package amazon_product_metadata_parser.output;
 
 import org.junit.Test;
 
+import amazon_product_metadata_parser.output.SQLServer2012.SQLServer2012Output;
+
 public class SQLServer2012OutputTest {
 
   @Test
